Visualize Objective

This class is used to create visualizations of optimization traces.

Value

Returns self invisibly.

Super classes

vistool::Visualizer -> vistool::Visualizer1D -> Visualizer1DObj

Public fields

objective

(`Objective`)
The objective function.

Methods

Public methods

Visualizer1DObj$new()

Visualizer1DObj$add_optimization_trace()

Visualizer1DObj$plot()

Visualizer1DObj$clone()

Inherited methods

vistool::Visualizer$add_points()

vistool::Visualizer$initialize_defaults()

vistool::Visualizer$save()

Method new()

Creates a new instance of this [R6][R6::R6Class] class.

Usage

Visualizer1DObj$new(objective, xlim = NULL, n_points = 100L)

Arguments

objective

(`Objective`)
The objective to optimize.

xlim

(`numeric(2)`)
Limits for the x-axis. If NULL, will be determined from objective bounds.

n_points

(`integer(1)`)
The number of generated point per dimension. Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.

n_points

(`integer(1)`)
The number of generated point per dimension. Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.

Method add_optimization_trace()

Add optimization trace to the plot.

Usage

Visualizer1DObj$add_optimization_trace(optimizer)

Arguments

optimizer

(`Optimizer`)
The optimizer to add to the plot.

Method plot()

Create and return the ggplot2 plot with optimization traces.

Usage

Visualizer1DObj$plot(...)

Arguments

...

Additional arguments passed to the parent plot method.

Returns

A ggplot2 object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Visualizer1DObj$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.