Visualize Objective

This class is used to create visualizations and animations of optimization traces.

Super classes

vistool::Visualizer -> vistool::Visualizer2D -> Visualizer2DObj

Public fields

objective

(`Objective`)
The objective function.

Methods

Public methods

Visualizer2DObj$new()

Visualizer2DObj$add_optimization_trace()

Visualizer2DObj$plot()

Visualizer2DObj$clone()

Inherited methods

vistool::Visualizer$add_points()

vistool::Visualizer$initialize_defaults()

vistool::Visualizer$save()

Method new()

Creates a new instance of this [R6][R6::R6Class] class.

Usage

Visualizer2DObj$new(
  objective,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L
)

Arguments

objective

(`Objective`)
The objective to optimize.

x1_limits

(`numeric(2)`)
The x1 limits.

x2_limits

(`numeric(2)`)
The x2 limits.

padding

(`numeric(1)`)
A margin that is added to x1limits and x2limits. The x1 margin is calculated by `max(x1lmits) - min(x1limits) * padding`.

n_points

(`integer(1)`)
The number of generated point per dimension. Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.

Method add_optimization_trace()

Add optimization trace to the plot.

Usage

Visualizer2DObj$add_optimization_trace(
  optimizer,
  line_color = NULL,
  line_width = 1.2,
  line_type = "solid",
  npoints = NULL,
  npmax = NULL,
  name = NULL,
  add_marker_at = 1,
  marker_size = 3,
  marker_shape = 16,
  marker_color = NULL,
  show_start_end = TRUE,
  alpha = 0.8
)

Arguments

optimizer

(`Optimizer`)
The optimizer from which the archive is extracted and used to plot the trace.

line_color

(`character(1)`)
The color of the trace line. If NULL, a random color is generated.

line_width

(`numeric(1)`)
The width of the trace line. Default is 1.2.

line_type

(`character(1)`)
The line type for the trace. Default is "solid".

npoints

(`integer(1)`)
The number of used points from the archive. Default is NULL which means all points are used.

npmax

(`integer(1)`)
The maximum number of points to use. Default is NULL.

name

(`character(1)`)
The name of the trace in the legend. Default is NULL which creates a name from optimizer and objective IDs.

add_marker_at

(`integer()`)
Vector of iterations at which markers are added. Default is first iteration only.

marker_size

(`numeric(1)`)
Size of the markers. Default is 3.

marker_shape

(`character(1)` or `integer(1)`)
Shape of the markers. Default is "circle" (16).

marker_color

(`character(1)`)
Color of the markers. If NULL, uses line_color.

show_start_end

(`logical(1)`)
Whether to highlight start and end points differently. Default is TRUE.

alpha

(`numeric(1)`)
Transparency level for the trace. Default is 0.8.

Returns

Returns self invisibly for method chaining.

Method plot()

Create and return the ggplot2 plot with optimization traces.

Usage

Visualizer2DObj$plot(...)

Arguments

...

Additional arguments passed to the parent plot method.

text_size

(`numeric(1)`)
Base text size for plot elements. Default is 11.

theme

(`character(1)`)
ggplot2 theme to use. One of "minimal", "bw", "classic", "gray", "light", "dark", "void". Default is "minimal".

Returns

A ggplot2 object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Visualizer2DObj$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.