Assertion for the main signature of a step_size_controlXX function.

Assertion for the main signature of a step_size_controlXX function.

Usage

assert_step_size_control(x, u, obj, opt)

Arguments

x

(numeric()) The "old" point.

u

(numeric()) The update added to x without a step size control.

obj

(Objective) The usd objective object.

opt

(Optimizer) The optimizer object from which the function is called.