Retrieve objective functions

Retrieve an objective function from the dictionary.

Usage

obj(.key, ...)

Arguments

.key

(character(1))
Key passed to the respective dictionary to retrieve the object.

...

(named list())
Named arguments passed to the constructor, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.