Visualize Base Class

This class is used to create 1D visualizations.

Public fields

x

(`vector()`)
x-values.

y

(`vector()`)
y-values

plot_lab

(character(1)
Label of the plot.

x_lab

(character(1)
Label of the x axis.

y_lab

(character(1)
Label of the y axis.

Methods

Public methods

Visualizer1D$new()

Visualizer1D$init_layer_lines()

Visualizer1D$setLayout()

Visualizer1D$plot()

Visualizer1D$save()

Visualizer1D$clone()

Method new()

Creates a new instance of this [R6][R6::R6Class] class.

Usage

Visualizer1D$new(x, y, plot_lab = NULL, x_lab = "x", y_lab = "y")

Arguments

x

(`numeric()`)
x-values.

y

(`numeric()`)
y-values.

plot_lab

(character(1)
Label of the plot.

x_lab

(character(1)
Label of the x axis.

y_lab

(character(1)
Label of the y axis.

Method init_layer_lines()

Initialize the plot with a line plot.

Usage

Visualizer1D$init_layer_lines(...)

Arguments

...

(`any`)
Further arguments passed to `add_trace(...)`.

Method setLayout()

Set the layout of the plotly plot.

Usage

Visualizer1D$setLayout(...)

Arguments

...

(`any`)
Layout options directly passed to `layout(...)`.

Method plot()

Return the plot and hence plot it or do further processing.

Usage

Visualizer1D$plot()

Method save()

Save the plot by using plotlys `orca()` function.

Usage

Visualizer1D$save(...)

Arguments

...

(`any`)
Further arguments passed to `orca()`.

Method clone()

The objects of this class are cloneable with this method.

Usage

Visualizer1D$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.