Visualizer for Losses

Visualize one or multiple loss functions.

Value

Returns self invisibly.

Super class

vistool::Visualizer -> VisualizerLossFuns

Public fields

losses

(`list`)
List of LossFunction objects.

task_type

(`character(1)`)
Task type (regr or classif).

lab_x

(`character(1)`)
Label for x-axis.

lab_y

(`character(1)`)
Label for y-axis.

title

(`character(1)`)
Title of the plot.

x_range

(`numeric(2)`)
Range for x-axis values.

line_width

(`numeric()`)
Line widths for different loss functions.

line_col

(`character()`)
Line colors for different loss functions.

line_type

(`character()`)
Line types for different loss functions.

legend_title

(`element`)
Legend title element.

y_pred

(`numeric()`)
Predicted values.

y_true

(`numeric()`)
True values.

n_points

(`integer(1)`)
Number of points to use for plotting the loss functions.

Methods

Public methods

VisualizerLossFuns$new()

VisualizerLossFuns$init_layer_lines()

VisualizerLossFuns$plot()

VisualizerLossFuns$clone()

Inherited methods

vistool::Visualizer$save()

Method new()

Creates a new instance of this [R6][R6::R6Class] class.

Usage

VisualizerLossFuns$new(losses, y_pred = NULL, y_true = NULL, n_points = 1000L)

Arguments

losses

(`list`)
List of LossFunction objects.

y_pred

(`numeric()`)
Predicted values. Optional.

y_true

(`numeric()`)
True values. Optional.

n_points

(`integer(1)`)
Number of points to use for plotting the loss functions. Default is 1000.

Method init_layer_lines()

Initialize line layer for the plot.

Usage

VisualizerLossFuns$init_layer_lines(color = NULL)

Arguments

color

(`character()`)
Colors to use for the lines. If NULL, default colors will be used.

Method plot()

Create and return the ggplot2 plot.

Usage

VisualizerLossFuns$plot(text_size = 11, theme = "bw")

Arguments

text_size

(`numeric(1)`)
Base text size for plot elements. Default is 11.

theme

(`character(1)`)
ggplot2 theme to use. One of "minimal", "bw", "classic", "gray", "light", "dark", "void". Default is "bw".

Returns

A ggplot2 object showing the loss functions.

Method clone()

The objects of this class are cloneable with this method.

Usage

VisualizerLossFuns$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.